+++ src/test/java/org/logstash/outputs/s3/GzipRecoverTest.java
@@ -1,0 +1,4 @@
+package test.java.org.logstash.outputs.s3;
+
+public class GzipRecoverTest {
+}